Refactor tests

+++ client/data_transfer.go
@@ -257,7 +257,7 @@ func (cli *Client) RequestDataQueryResponseFromVerifiers(ctx context.Context, da
 }
 
 // DownloadFile requests a file download from file hoster given the file segments start and end.
-func (cli *Client) DownloadFile(ctx context.Context, contractHash, fileHash string, fileSize uint64, from, to int32) (string, error) {
+func (cli *Client) DownloadFile(ctx context.Context, contractHash, fileHash string, fileSize uint64, from, to int64) (string, error) {
 	payload := JSONRPCRequest{
 		JSONRPC: "2.0",
 		Method:  "data_transfer.DownloadFile",
